check_in_gui/pyproject.toml
[tool.poetry]
name = "check_in_gui"
version = "0.1.0"
description = ""
authors = ["= <doublenegative@protonmail.com>"]

[tool.poetry.dependencies]
python = "3.9"
Kivy = "^2.0.0"
matplotlib = "^3.4.3"
requests = "^2.26.0"
grpcio = "^1.41.1"
grpcio-tools = "^1.41.1"
toml = "^0.10.2"
click = "^8.0.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pyinstaller = "^4.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
